re-integrate support for laminar APA generation

+++ mktasks.py
@@ -5,6 +5,8 @@ from __future__ import division
 import argparse
 import sys
 
+from math import ceil
+
 from os.path import exists, dirname
 from os import makedirs
 
@@ -21,10 +23,7 @@ def is_feasible(taskset):
     aff = get_native_affinities(taskset)
     ts  = get_native_taskset(taskset)
     sol = apa_implicit_deadline_feasible(ts, aff)
-    if sol:
-        return True
-    else:
-        return False
+    return True if sol else False
 
 
 def make_taskset(n, u, min_wcet=500):
@@ -39,9 +38,45 @@ def make_taskset(n, u, min_wcet=500):
 
     return ts
 
+def three_level_affinities(m, num_sockets):
+    per_socket = int(ceil(m / num_sockets))
+    all_cores = frozenset(range(0, m))
+
+    sockets = []
+    for i in range(0, num_sockets):
+        s = range(i * per_socket, (i + 1) * per_socket)
+        # check for last incomplete socket
+        while s[-1] >= m:
+            del s[-1]
+        sockets.append(frozenset(s))
+    parts = [frozenset([x]) for x in all_cores]
+    return ([all_cores], sockets, parts)
+
+
+def assign_three_level_affinities(ts, m, sockets, max_tries=10):
+    affinities = three_level_affinities(m, sockets)
+
+    # initially global
+    for t in ts:
+        t.affinity = affinities[0][0]
+
+    assert is_feasible(ts)
+
+    for t in ts:
+        attempts = 1
+        while True:
+            attempts += 1
+            group = random.choice(affinities)
+            t.affinity = random.choice(group)
+            if is_feasible(ts):
+                break
+            if attempts >= max_tries:
+                # restore global
+                t.affinity = affinities[0][0]
+                break
 
 def all_possible_affinities(m):
-    all_cores = frozenset(range(1, m))
+    all_cores = frozenset(range(0, m))
     to_look_at = [all_cores]
     while to_look_at:
         aff = to_look_at.pop()
@@ -51,14 +86,11 @@ def all_possible_affinities(m):
             mid = sorted(aff)[len(aff) // 2]
             left  = frozenset([x for x in aff if x < mid])
             right = frozenset([x for x in aff if x >= mid])
-#            print aff, mid, left, right
             to_look_at.append(left)
             to_look_at.append(right)
 
 def assign_random_laminar_affinities(ts, m, max_tries=10):
     all_picks = list(all_possible_affinities(m))
-
-    print all_picks[0], ts.utilization()
 
     # initially global
     for t in ts:
@@ -71,7 +103,6 @@ def assign_random_laminar_affinities(ts, m, max_tries=10):
         while True:
             attempts += 1
             t.affinity = random.choice(all_picks)
-#             print t, 'trying', t.affinity
             if is_feasible(ts):
                 break
             if attempts >= max_tries:
@@ -80,6 +111,7 @@ def assign_random_laminar_affinities(ts, m, max_tries=10):
                 break
 
 def assign_random_priorities(ts):
+    "assign random priorities"
     prios = range(1, len(ts) + 1)
     random.shuffle(prios)
     for (t, p) in zip(ts, prios):
@@ -87,11 +119,13 @@ def assign_random_priorities(ts):
     ts.assign_ids()
 
 def assign_rm_priorities(ts):
+    "assign rate-monotomic priorities"
     ts.assign_ids_by_period()
     for t in ts:
         t.priority = t.id
 
 def assign_arm_priorities(ts):
+    "assign affinity- and rate-monotonic priorities"
     for (t, i) in zip(sorted(ts, key=lambda t: (1/len(t.affinity), t.period)), range(1, len(ts) + 1)):
         t.priority = i
     ts.assign_ids()
@@ -109,7 +143,6 @@ def to_json(ts):
             'cost'     : t.cost,
             'period'   : t.period,
             'affinity' : list(t.affinity),
-#             'affinity_mask' : to_hex(t.affinity),
             'priority' : t.priority,
         }
     data = {
@@ -117,12 +150,13 @@ def to_json(ts):
     }
     return json.dumps(data, sort_keys=True, indent=4, separators=(',', ': '))
 
-def store_taskset(m, n, u, seq, prefix=''):
-    print u * (m - 1)
-    ts = make_taskset(n, u * (m - 1))
+def store_random_taskset(m, n, u, seq, prefix=''):
+    print "[random laminar APAs, %d cores, %.2f utilization, and %d tasks per core]" \
+             % (m, u, n)
+    ts = make_taskset(n, u * m)
     assign_random_laminar_affinities(ts, m)
     assign_arm_priorities(ts)
-    fname = "%sapa-workload_m=%02d_n=%02d_u=%2d_seq=%02d.json" % \
+    fname = "%sapa-r-workload_m=%02d_n=%02d_u=%2d_seq=%02d.json" % \
         (prefix, m, n, int(100 * u), seq)
     print '=>', fname
     f  = open(fname, 'w')
@@ -140,9 +174,27 @@ def store_partitioned_taskset(m, n, u, seq, prefix=''):
             t.affinity = set([core])
         ts += per_core
 
-    assign_arm_priorities(ts)
+    assign_rm_priorities(ts)
     fname = "%spart-workload_m=%02d_n=%02d_u=%2d_seq=%02d.json" % \
         (prefix, m, n * m, int(100 * u), seq)
+    print '=>', fname
+    f  = open(fname, 'w')
+    f.write(to_json(ts))
+    f.close()
+
+def store_socket_taskset(m, sockets, n, u, seq, prefix=''):
+    print "[socket-aware laminar APAs, %d cores, %d sockets, %.2f utilization, and %d tasks per core]" \
+         % (m, sockets, u, n)
+    fname = "%sapa-s-workload_m=%02d_s=%02d_n=%02d_u=%2d_seq=%02d.json" % \
+        (prefix, m, sockets, n, int(100 * u), seq)
+    if exists(fname):
+        print 'Skipping %s; exists already.' % fname
+        return
+    ts = make_taskset(n, u)
+    assign_three_level_affinities(ts, m, sockets)
+    assign_rm_priorities(ts)
+    fname = "%sapa-workload_m=%02d_n=%02d_u=%2d_seq=%02d.json" % \
+        (prefix, m, n, int(100 * u), seq)
     print '=>', fname
     f  = open(fname, 'w')
     f.write(to_json(ts))
@@ -158,8 +210,11 @@ def parse_args():
         help='Prefix for the generated file[s]')
 
     p.add_argument(
-        '-m', '--num-cores', type=int, nargs='*', dest='ncores', default=[1],
+        '-m', '--num-cores', type=int, nargs='*', dest='ncores', default=[],
         help='processor counts to consider [multiple possible]')
+    p.add_argument(
+        '-s', '--num-sockets', type=int, nargs='*', dest='nsockets', default=[1],
+        help='socket counts to consider [multiple possible]')
     p.add_argument(
         '-n', '--tasks-per-core', type=int, nargs='*', dest='ntasks', default=[5],
         help='task counts to consider [multiple possible]')
@@ -167,6 +222,11 @@ def parse_args():
         '-u', '--per-core-utilization', type=float, nargs='*', dest='utils',
         default=[0.5],
         help='processor utilizations to consider [multiple possible]')
+
+    p.add_argument(
+        '--apa', type=str, choices=['partitioned', 'random', 'socket'],
+            dest='apa_type', default='partitioned',
+        help='what sort of affinities to generate [default: partitioned]')
 
     p.add_argument(
         '-c', '--count', type=int, dest='count', default=1,
@@ -185,8 +245,18 @@ def main(args=sys.argv[1:]):
         for n in options.ntasks:
             for u in options.utils:
                 for seq in xrange(options.count):
-                    store_partitioned_taskset(m, n, u, seq,
-                        prefix=options.prefix)
+                    if options.apa_type == 'partitioned':
+                        store_partitioned_taskset(m, n, u, seq,
+                            prefix=options.prefix)
+                    elif options.apa_type == 'random':
+                        store_random_taskset(m, n, u, seq,
+                            prefix=options.prefix)
+                    elif options.apa_type == 'socket':
+                        for s in options.nsockets:
+                            store_socket_taskset(m, s, n, u, seq,
+                                prefix=options.prefix)
+                    else:
+                        assert False
 
 if __name__ == '__main__':
     main()
